+++ controllers/ControllerProduto.js
@@ -2,8 +2,8 @@ var db = require("../config/conexao")
 const Produto = require("../models/Produto")
 const Modelo = require("../models/Modelo")
 const Fabricante = require("../models/Fabricante")
-const Venda = require("../models/Venda")
 const ItensVenda = require("../models/ItensVenda")
+const ItensCompra = require("../models/ItensCompra")
 const Empresa = require("../models/Empresa")
 
 exports.listAll = (req, res) => {
@@ -15,12 +15,6 @@ exports.listAll = (req, res) => {
 	}).then((dadosProduto) => {
 		Modelo.findAll({where: {ativo: 'Ativo'}}).then((dadosModelo) => {
 			Fabricante.findAll({where: {ativo: 'Ativo'}}).then((dadosFabricante) => {
-
-					//SELECT EXTRACT(DAY FROM "dataVenda") as "diaVenda", EXTRACT(MONTH FROM "dataVenda") as "mesVenda" FROM "vendas"
-					//SELECT "produtoId" AS "id", SUM(quantidade) AS "demanda", EXTRACT(MONTH FROM "createdAt") AS "mes" FROM "itens_vendas" GROUP BY  EXTRACT(MONTH FROM "createdAt"), "produtoId"
-					//SELECT SUM(iv."quantidade") AS "demanda", iv."produtoId" AS "id", p."prazoReposicao" AS "tempoDeEntrega", EXTRACT(MONTH FROM v."dataVenda") AS "mes" FROM "itens_vendas" iv INNER JOIN "produtos" p on p."id" = iv."produtoId" INNER JOIN "vendas" v on v."id" = iv."vendaId" GROUP BY EXTRACT(MONTH FROM v."dataVenda"), iv."produtoId"
-					//'WHERE v."dataVenda" BETWEEN NOW() - INTERVAL 30 DAY AND NOW()' +
-
 					const contextProduto = {
 						produtos: dadosProduto.map(dado => {
 							return {
@@ -91,23 +85,29 @@ exports.add = async (req, res) => {
 	res.redirect("/produtos/list-produtos")
 }
 
-exports.delete = (req, res) => {
-	ItensVenda.findAll({where: {produtoId: req.body.id}}).then((itensVenda) =>{
-		if(itensVenda.length < 1){
-			Produto.destroy({where: {id: req.body.id}}).then(() => {
-				req.flash("msg_sucesso", "Produto deletado com sucesso!")
-				res.redirect("/produtos/list-produtos")
-			}).catch((erro) => {
-				req.flash("msg_erro", "Não foi possivel deletar o produto. Erro: " + erro)
-				res.redirect("/produtos/list-produtos")
-			})
-		}else {
-			req.flash("msg_erro", "Não é possivel excluir esse Produto, pois esta sendo utilizado em uma Venda!")
+exports.delete = async (req, res) => {
+	const itensVenda = await ItensVenda.findAll({where: {produtoId: req.body.id}})
+	const itensCompra = await ItensCompra.findAll({where: {produtoId: req.body.id}})
+	if(itensVenda.length < 1 && itensCompra.length < 1){
+		Produto.destroy({where: {id: req.body.id}}).then(() => {
+			req.flash("msg_sucesso", "Produto deletado com sucesso!")
 			res.redirect("/produtos/list-produtos")
-		}
-	}).catch((erro)=>{
-		req.flash("msg_erro", "Não foi possivel encontrar o produto! " + erro)
+		}).catch((erro) => {
+			req.flash("msg_erro", "Não foi possivel deletar o produto. Erro: " + erro)
+			res.redirect("/produtos/list-produtos")
+		})
+	}else {
+		req.flash("msg_erro", "O produto está sendo utilizado em um registro")
 		res.redirect("/produtos/list-produtos")
+	}
+
+
+
+	Pessoa.findByPk(req.body.id).then((pessoa) => {
+
+	}).catch((erro) => {
+		req.flash("msg_erro", "Não foi possivel encontrar o pessoa! " + erro)
+		res.redirect("/pessoas/list-pessoas")
 	})
 }
